# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "vision-api (face)",
...
        "contact": {
            "name": "vision-stack",
            "url": "http://192.168.1.50:18090/docs"
        },
        "license": {
            "name": "internal"
        },
        "version": "1.1.0"
    },
    "paths": {
...
        "/v1/faces/detect": {
            "post": {
                "tags": [
                    "faces"
                ],
                "summary": "Ph\u00e1t hi\u1ec7n khu\u00f4n m\u1eb7t (bbox, kh\u00f4ng nh\u1eadn d\u1ea1ng)",
                "operationId": "faces_detect_v1_faces_detect_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "x-tenant-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "B\u1eaft bu\u1ed9c khi token l\u00e0 global '*'.",
                            "default": "",
                            "title": "X-Tenant-Id"
                        },
                        "description": "B\u1eaft bu\u1ed9c khi token l\u00e0 global '*'."
                    },
                    {
                        "name": "x-request-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "T\u00f9y ch\u1ecdn; echo l\u1ea1i trong response.",
                            "default": "",
                            "title": "X-Request-Id"
                        },
                        "description": "T\u00f9y ch\u1ecdn; echo l\u1ea1i trong response."
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_faces_detect_v1_faces_detect_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetectResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\u1ea2nh/URL kh\u00f4ng h\u1ee3p l\u1ec7",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Thi\u1ebfu / sai token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sai tenant ho\u1eb7c thi\u1ebfu quy\u1ec1n admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "V\u01b0\u1ee3t rate limit c\u1ee7a tenant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service ch\u01b0a `ready`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/faces/embed": {
            "post": {
                "tags": [
                    "faces"
                ],
                "summary": "Tr\u00edch embedding 512-d cho m\u1ed7i khu\u00f4n m\u1eb7t (\u0111\u1ec3 enroll)",
                "description": "K\u1ebft qu\u1ea3 `faces[i].embedding` (512 s\u1ed1, \u0111\u00e3 L2-norm) l\u00e0 d\u1eef li\u1ec7u g\u1eedi v\u00e0o backend khi enroll.",
                "operationId": "faces_embed_v1_faces_embed_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "x-tenant-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "B\u1eaft bu\u1ed9c khi token l\u00e0 global '*'.",
                            "default": "",
                            "title": "X-Tenant-Id"
                        },
                        "description": "B\u1eaft bu\u1ed9c khi token l\u00e0 global '*'."
                    },
                    {
                        "name": "x-request-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "T\u00f9y ch\u1ecdn; echo l\u1ea1i trong response.",
                            "default": "",
                            "title": "X-Request-Id"
                        },
                        "description": "T\u00f9y ch\u1ecdn; echo l\u1ea1i trong response."
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_faces_embed_v1_faces_embed_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmbedResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\u1ea2nh/URL kh\u00f4ng h\u1ee3p l\u1ec7",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Thi\u1ebfu / sai token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sai tenant ho\u1eb7c thi\u1ebfu quy\u1ec1n admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "V\u01b0\u1ee3t rate limit c\u1ee7a tenant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service ch\u01b0a `ready`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/index/stats": {
            "get": {
                "tags": [
                    "faces"
                ],
                "summary": "S\u1ed1 person / vector \u0111\u00e3 index",
                "operationId": "index_stats_v1_index_stats_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "x-tenant-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "B\u1eaft bu\u1ed9c khi token l\u00e0 global '*'.",
                            "default": "",
                            "title": "X-Tenant-Id"
                        },
                        "description": "B\u1eaft bu\u1ed9c khi token l\u00e0 global '*'."
                    },
                    {
                        "name": "x-request-id",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "T\u00f9y ch\u1ecdn; echo l\u1ea1i trong response.",
                            "default": "",
                            "title": "X-Request-Id"
                        },
                        "description": "T\u00f9y ch\u1ecdn; echo l\u1ea1i trong response."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\u1ea2nh/URL kh\u00f4ng h\u1ee3p l\u1ec7",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Thi\u1ebfu / sai token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sai tenant ho\u1eb7c thi\u1ebfu quy\u1ec1n admin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "V\u01b0\u1ee3t rate limit c\u1ee7a tenant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service ch\u01b0a `ready`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_faces_detect_v1_faces_detect_post": {
                "properties": {
                    "file": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "binary"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "File",
                        "description": "\u1ea2nh upload (jpg/png/\u2026)"
                    },
                    "url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Url",
                        "description": "Ho\u1eb7c URL \u1ea3nh http/https"
                    }
                },
                "type": "object",
                "title": "Body_faces_detect_v1_faces_detect_post"
            },
...
            "DetectResponse": {
                "properties": {
                    "request_id": {
                        "type": "string",
                        "title": "Request Id",
                        "examples": [
                            "req_f6c383fddaf74d09"
                        ]
                    },
                    "tenant_id": {
                        "type": "string",
                        "title": "Tenant Id",
                        "examples": [
                            "t_demo"
                        ]
                    },
                    "count": {
                        "type": "integer",
                        "title": "Count",
                        "description": "S\u1ed1 khu\u00f4n m\u1eb7t t\u00ecm th\u1ea5y.",
                        "examples": [
                            2
                        ]
                    },
                    "faces": {
                        "items": {
                            "$ref": "#/components/schemas/FaceBox"
                        },
                        "type": "array",
                        "title": "Faces"
                    },
                    "inference_ms": {
                        "type": "number",
                        "title": "Inference Ms",
                        "description": "Th\u1eddi gian suy lu\u1eadn GPU (ms).",
                        "examples": [
                            41.2
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "request_id",
                    "tenant_id",
                    "count",
                    "faces",
                    "inference_ms"
                ],
                "title": "DetectResponse"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "description": "API token c\u1ea5p cho 1 tenant, ho\u1eb7c token global (role=admin, tenant='*').",
                "scheme": "bearer"
            }
        }
    },
...
}